MNT: use new 94-load.py
The new version xpdacq uses the `UserInterface` class to setup the objects in the ipython session. This enables future changes to the setup to happen in xpdacq package and aovid frequent changing in profile_collection. The new file also cleans up the name space and only exposre necesary objects to users.

The old file is changed to 94-load.bak.